chore: adopt cargo fmt + fix clippy warnings
Adopt cargo fmt as the project formatter (replaces the prior
hand-formatted wide-line style) and clear the clippy lints.

- cargo fmt across 12 files (batch, container, engine, html, md,
  translate, ...): pure reformatting, no behavior change.
- clippy: drop an unnecessary `as usize` cast (archive.rs), remove
  `let _ =` bindings on unit-returning calls (html.rs x2), and use
  `contains()` over `iter().any()` in two ass.rs tests.

Verified: clippy --all-targets = 0 warnings; 76 tests pass.

+++ src/archive.rs
@@ -23,7 +23,7 @@ impl Epub {
     pub fn load(path: &Path) -> Result<Self> {
         let file = File::open(path).with_context(|| format!("open {}", path.display()))?;
         let mut za = ZipArchive::new(file).context("read zip")?;
-        let mut entries = Vec::with_capacity(za.len() as usize);
+        let mut entries = Vec::with_capacity(za.len());
         for i in 0..za.len() {
             let mut zf = za.by_index(i)?;
             let name = zf.name().to_string();
